requirements.txt
Flask
tkinterweb
dominate
mysql-connector-python
Pillow